- Adding a print ZIF state function
- Splitting read(..) to read_byte(..). read(..) merely takes a range
argument and calls read_byte(..) in a loop. The intention is to later
split off initialization routines from the loop, and to allow bytes to
be returned from read_byte(..) for other functions.

- Related to above, fixed a bug where ZIF state wouldn't reset before
incrementing addresses...

- Updated help text. Note that I'm not accepting hex yet, even though it
says hex in the help text -- that will be in a commit or two.

- Implemented a prototype for driving the clock pin directly, but
doesn't currently appear to work. Will continue this once I have access
to a logic analyzer.

- Implemented the lock routine (and several helper functions), but it
does not currently work. Once I have access to a logic analyzer and
verify that other programmers can set the lockbit on my target chip,
I'll continue work on this.

- Implemented a test function, but it does not yet compare results for
autonomous checking.

- Implemented a blank check.